lf-6.6.52-2.2.0->6.8 rebase-fix: cummulatice fixes after rebase
Cummulative fixes picked after clean rebase to 6.6.23

+++ drivers/firmware/arm_scmi/perf.c
@@ -155,7 +155,7 @@ struct perf_dom_info {
 	u32 opp_count;
 	u32 sustained_freq_khz;
 	u32 sustained_perf_level;
-	unsigned long mult_factor;
+	u32 mult_factor;
 	struct scmi_perf_domain_info info;
 	struct scmi_opp opp[MAX_OPPS];
 	struct scmi_fc_info *fc_info;
